pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "storify"
version = "0.1.0"
description = "A simple fastapi based server for storing and distributing files"
readme = "README.md"
requires-python = ">=3.8,<3.13"
license = { text = "Apache-2.0" }
authors = [
    { name = "Arian Ott", email = "arian.ott@ieee.org" }
]
keywords = ["cli", "stories", "snippets", "organiser"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent"
]
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "pydantic-settings",
    "python-multipart",
    "python-jose",
    "argon2-cffi",
    "passlib",
    "pymysql",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "pytest-xdist",
]

[project.scripts]
storify = "storify.cli:app"

[project.urls]
Homepage = "https://github.com/arian-ott/storify"
Repository = "https://github.com/arian-ott/storify"
Issues = "https://github.com/arian-ott/storify/issues"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-n auto"